Move UI message handlers to router module
